Add files via upload
The data includes 9 separate datasets that are linked with a primary key. This is a large dataset that covers several facets of the Brazilian e-commerce industry. It contains information on a variety of businesses that are currently operating on the market. It contains information about 100k orders scheduled from 2016 to 2018